+++ cmd/exp.go
@@ -13,9 +13,24 @@ import (
 	"github.com/ysicing/ergo/common"
 	"github.com/ysicing/ergo/pkg/config"
 	"github.com/ysicing/ergo/pkg/ergo/experimental"
+	"github.com/ysicing/ergo/pkg/ergo/experimental/codegen"
 	"github.com/ysicing/ergo/pkg/util/factory"
 	"helm.sh/helm/v3/pkg/cli/output"
 )
+
+func newCodeGenCmd(f factory.Factory) *cobra.Command {
+	c := &codegen.CodeOptions{
+		Log: f.GetLog(),
+	}
+	cmd := &cobra.Command{
+		Use:   "code [flags]",
+		Short: "初始化项目",
+		Run: func(cobraCmd *cobra.Command, args []string) {
+			c.Init()
+		},
+	}
+	return cmd
+}
 
 func newConfigCmd() *cobra.Command {
 	var configCommand = &cobra.Command{
@@ -107,5 +122,6 @@ func newExperimentalCmd(f factory.Factory) *cobra.Command {
 	cmd.AddCommand(install)
 	cmd.AddCommand(simplefile)
 	cmd.AddCommand(newConfigCmd())
+	cmd.AddCommand(newCodeGenCmd(f))
 	return cmd
 }