+++ pkg/ergo/experimental/codegen/code.go
@@ -1,40 +1,21 @@
-/*
- * Copyright (c) 2021 ysicing <i@ysicing.me>
- */
-
-package cmd
+// AGPL License
+// Copyright (c) 2022 ysicing <i@ysicing.me>
+package codegen
 
 import (
 	"strings"
 
 	"github.com/ergoapi/log"
 	"github.com/manifoldco/promptui"
-	"github.com/spf13/cobra"
-	"github.com/ysicing/ergo/pkg/ergo/codegen"
-	"github.com/ysicing/ergo/pkg/util/factory"
 )
 
 type CodeOptions struct {
 	Log log.Logger
 }
 
-func newCodeGenCmd(f factory.Factory) *cobra.Command {
-	c := &CodeOptions{
-		Log: f.GetLog(),
-	}
-	cmd := &cobra.Command{
-		Use:   "code [flags]",
-		Short: "初始化项目",
-		Run: func(cobraCmd *cobra.Command, args []string) {
-			c.Init()
-		},
-	}
-	return cmd
-}
-
 func (code CodeOptions) Init() {
 	searcher := func(input string, index int) bool {
-		p := codegen.CodeType[index]
+		p := CodeType[index]
 		name := strings.Replace(strings.ToLower(p.Key), " ", "", -1)
 		input = strings.Replace(strings.ToLower(input), " ", "", -1)
 		return strings.Contains(name, input)
@@ -47,15 +28,15 @@ func (code CodeOptions) Init() {
 	}
 	codetype := promptui.Select{
 		Label:     "选择代码类型",
-		Items:     codegen.CodeType,
+		Items:     CodeType,
 		Searcher:  searcher,
 		Size:      4,
 		Templates: templates,
 	}
 	codetypeid, _, _ := codetype.Run()
-	selectcodetypevalue := codegen.CodeType[codetypeid].Key
+	selectcodetypevalue := CodeType[codetypeid].Key
 	code.Log.Infof("\U0001F389 选择 %v", selectcodetypevalue)
-	codefunc := codegen.CodeGen{Log: code.Log}
+	codefunc := CodeGen{Log: code.Log}
 	if selectcodetypevalue == "go" {
 		code.Log.Infof("Start downloading the template...")
 		if err := codefunc.GoClone(); err != nil {